requirements.txt
flask==2.3.2
flask-cors==4.0.0
gunicorn
numpy==1.25.0
pymongo==4.4.0
Pillow==9.5.0
python-dotenv==1.0.0

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1

--extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.15.2